Revert "itemの画像を表示するところまでできたが、URLをハードコピーしているので改善したい"
This reverts commit 5f5dfe7fdc4f10e1c765c9f4dc0391458a1df3df.
